@elazarl 2019-04-10
Thanks, looks clean and reasonable.

I'd love to see another PR of an.in memory cache