@suxianbaozi 2021-11-12
> > how to use this...
> 
> Haven't used it, but this test looks like a good example: [9718da4#diff-1ad863f10fb8c2c1ed4cbd0c2316a75ece3f5c78458506ca8c95b5f03533236aR800](https://github.com/elazarl/goproxy/commit/9718da4410e558da9da518a6286e25cb4442d3d0#diff-1ad863f10fb8c2c1ed4cbd0c2316a75ece3f5c78458506ca8c95b5f03533236aR800)

thanks a lot